Finished The Second Tutorial
I followed the second tutorial and created an interactable object.  In the current the state the player will view the object as an interactable and move towards it then the object is right clicked.  The player will stop some distance from it.  The space for the interactactable object can be resized and moved.